# apps/api/contract/openapi.yaml
openapi: 3.1.0
info:
  title: LearnR API
  version: 0.1.0
...
paths:
...
  /me:
    get:
      operationId: readAccount
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  role:
                    anyOf:
                      - type: string
                        enum:
                          - parent
                          - child
                      - type: 'null'
                  parentId:
                    anyOf:
                      - type: string
                      - type: 'null'
                  name:
                    anyOf:
                      - type: string
                      - type: 'null'
                  avatar:
                    anyOf:
                      - type: string
                        enum:
                          - fox
                          - bear
                          - cat
                          - owl
                          - frog
                          - whale
                          - rabbit
                          - panda
                      - type: 'null'
                  image:
                    anyOf:
                      - type: string
                      - type: 'null'
                  photo:
                    anyOf:
                      - type: string
                      - type: 'null'
                required:
                  - id
                  - role
                  - parentId
                  - name
                  - avatar
                  - image
                  - photo
                additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
...
  /speed/runs:
    post:
      operationId: submitSpeedRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                  pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                mode:
                  type: string
                  minLength: 1
                correct:
                  type: integer
                  minimum: 0
                  maximum: 10000
                playedAt:
                  type: string
                  format: date-time
                  description: >-
                    When the run was played, ISO 8601. Omit it and the server
                    stamps receipt. Hold one stamp per run across every flush of
                    it, as with `id`.
              required:
                - id
                - mode
                - correct
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  previousBest:
                    anyOf:
                      - type: integer
                        minimum: -9007199254740991
                        maximum: 9007199254740991
                      - type: 'null'
                  best:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  isRecord:
                    type: boolean
                  standing:
                    anyOf:
                      - type: object
                        properties:
                          place:
                            type: integer
                            minimum: -9007199254740991
                            maximum: 9007199254740991
                          previousPlace:
                            anyOf:
                              - type: integer
                                minimum: -9007199254740991
                                maximum: 9007199254740991
                              - type: 'null'
                          rivals:
                            type: integer
                            minimum: -9007199254740991
                            maximum: 9007199254740991
                        required:
                          - place
                          - previousPlace
                          - rivals
                        additionalProperties: false
                      - type: 'null'
                required:
                  - previousBest
                  - best
                  - isRecord
                  - standing
                additionalProperties: false
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /speed/records:
    get:
      operationId: readFamilyRecords
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempts:
                    type: array
                    items:
                      type: object
                      properties:
                        mode:
                          type: string
                        correct:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        playedAt:
                          type: string
                          format: date-time
                      required:
                        - mode
                        - correct
                        - playedAt
                      additionalProperties: false
                  family:
                    anyOf:
                      - type: array
                        items:
                          type: object
                          properties:
                            playerId:
                              type: string
                            playerName:
                              type: string
                            playerPhoto:
                              anyOf:
                                - type: string
                                - type: 'null'
                            playerAvatar:
                              anyOf:
                                - type: string
                                  enum:
                                    - fox
                                    - bear
                                    - cat
                                    - owl
                                    - frog
                                    - whale
                                    - rabbit
                                    - panda
                                - type: 'null'
                            playerImage:
                              anyOf:
                                - type: string
                                - type: 'null'
                            mode:
                              type: string
                            best:
                              type: integer
                              minimum: -9007199254740991
                              maximum: 9007199254740991
                            achievedAt:
                              type: string
                              format: date-time
                          required:
                            - playerId
                            - playerName
                            - mode
                            - best
                            - achievedAt
                          additionalProperties: false
                      - type: 'null'
                required:
                  - attempts
                  - family
                additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /speed/summaries:
    get:
      operationId: readSpeedSummaries
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    mode:
                      type: string
                    correct:
                      type: integer
                      minimum: -9007199254740991
                      maximum: 9007199254740991
                    playedAt:
                      type: string
                      format: date-time
                  required:
                    - mode
                    - correct
                    - playedAt
                  additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /speed/unseen:
    get:
      operationId: readUnseenRecords
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    childId:
                      type: string
                    childName:
                      type: string
                    mode:
                      type: string
                    best:
                      type: integer
                      minimum: -9007199254740991
                      maximum: 9007199254740991
                    achievedAt:
                      type: string
                      format: date-time
                  required:
                    - childId
                    - childName
                    - mode
                    - best
                    - achievedAt
                  additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /speed/unseen/{childId}:
    delete:
      operationId: dismissSpeedRecords
      parameters:
        - schema:
            type: string
          in: path
          name: childId
          required: true
      responses:
        '204':
          description: Default Response
...